Add `IViewModel.createLineBreaksComputer()`
